refactor/tests
Large refactor, remove mock_routing, default tests to use mock-crust

TBD doc warnings
