+++ src/personas/maid_manager.rs
@@ -56,7 +56,7 @@ impl Default for Account {
 }
 
 impl Account {
-    fn put_data(&mut self) -> Result<(), MutationError> {
+    fn add_entry(&mut self) -> Result<(), MutationError> {
         if self.space_available < 1 {
             return Err(MutationError::LowBalance);
         }
@@ -66,7 +66,7 @@ impl Account {
         Ok(())
     }
 
-    fn delete_data(&mut self) {
+    fn remove_entry(&mut self) {
         self.data_stored -= 1;
         self.space_available += 1;
         self.version += 1;
@@ -138,7 +138,7 @@ impl MaidManager {
             Some(client_request) => {
                 // Refund account
                 match self.accounts.get_mut(&utils::client_name(&client_request.src)) {
-                    Some(account) => account.delete_data(),
+                    Some(account) => account.remove_entry(),
                     None => return Ok(()),
                 }
                 let client_name = utils::client_name(&client_request.src);
@@ -276,7 +276,7 @@ impl MaidManager {
                          .get_mut(&client_name)
                          .ok_or(MutationError::NoSuchAccount)
                          .and_then(|account| {
-                             let result = account.put_data();
+                             let result = account.add_entry();
                              trace!("Client account {:?}: {:?}", client_name, account);
                              result
                          });
@@ -326,26 +326,19 @@ impl MaidManager {
 
 #[cfg(test)]
 #[cfg_attr(feature="clippy", allow(indexing_slicing))]
-#[cfg(not(feature="use-mock-crust"))]
+#[cfg(feature="use-mock-crust")]
 mod test {
     use super::*;
-    use super::Refresh;
-    use error::InternalError;
-    use safe_network_common::client_errors::MutationError;
-    use kademlia_routing_table::GROUP_SIZE;
-    use maidsafe_utilities::serialisation;
-    use rand::{thread_rng, random};
+    use test_utils;
+    use mock_crust_detail::{poll, test_node};
+    use mock_crust_detail::test_client::TestClient;
+    use rand::{random, thread_rng};
     use rand::distributions::{IndependentSample, Range};
-    use routing::{Authority, Data, ImmutableData, MessageId, RequestContent, RequestMessage,
-                  ResponseContent, StructuredData};
-    use sodiumoxide::crypto::hash::sha512;
-    use sodiumoxide::crypto::sign;
-    use std::rc::Rc;
-    use std::sync::mpsc;
-    use utils;
-    use utils::generate_random_vec_u8;
-    use vault::RoutingNode;
-    use xor_name::XorName;
+    use routing::{Data, ImmutableData};
+    use routing::mock_crust::{self, Network};
+    use kademlia_routing_table::GROUP_SIZE;
+    use mock_crust_detail;
+    const TEST_NET_SIZE: usize = GROUP_SIZE + 2; // just larger than CLOSE_GROUP
 
     #[test]
     fn account_ok() {
@@ -354,13 +347,13 @@ mod test {
         assert_eq!(0, account.data_stored);
         assert_eq!(super::DEFAULT_ACCOUNT_SIZE, account.space_available);
         for _ in 0..super::DEFAULT_ACCOUNT_SIZE {
-            assert!(account.put_data().is_ok());
+            assert!(account.add_entry().is_ok());
         }
         assert_eq!(super::DEFAULT_ACCOUNT_SIZE, account.data_stored);
         assert_eq!(0, account.space_available);
 
         for _ in 0..super::DEFAULT_ACCOUNT_SIZE {
-            account.delete_data();
+            account.remove_entry();
         }
         assert_eq!(0, account.data_stored);
         assert_eq!(super::DEFAULT_ACCOUNT_SIZE, account.space_available);
@@ -373,537 +366,123 @@ mod test {
         assert_eq!(0, account.data_stored);
         assert_eq!(super::DEFAULT_ACCOUNT_SIZE, account.space_available);
         for _ in 0..super::DEFAULT_ACCOUNT_SIZE {
-            assert!(account.put_data().is_ok());
+            assert!(account.add_entry().is_ok());
         }
         assert_eq!(super::DEFAULT_ACCOUNT_SIZE, account.data_stored);
         assert_eq!(0, account.space_available);
-        assert!(account.put_data().is_err());
+        assert!(account.add_entry().is_err());
         assert_eq!(super::DEFAULT_ACCOUNT_SIZE, account.data_stored);
         assert_eq!(0, account.space_available);
     }
-
-
-    struct Environment {
-        our_authority: Authority,
-        client: Authority,
-        routing: Rc<RoutingNode>,
-        maid_manager: MaidManager,
-    }
-
-    fn environment_setup() -> Environment {
-        let routing = unwrap_result!(RoutingNode::new(mpsc::channel().0, false));
-        let from = random::<XorName>();
-        let client;
-
-        loop {
-            let keys = sign::gen_keypair();
-            let name = XorName(sha512::hash(&keys.0[..]).0);
-            if let Ok(Some(_)) = routing.close_group(name) {
-                client = Authority::Client {
-                    client_key: keys.0,
-                    peer_id: random(),
-                    proxy_node_name: from,
-                };
-                break;
-            }
-        }
-
-        let routing = Rc::new(routing);
-
-        Environment {
-            our_authority: Authority::ClientManager(utils::client_name(&client)),
-            client: client,
-            routing: routing.clone(),
-            maid_manager: MaidManager::new(routing.clone()),
-        }
-    }
-
-    fn create_account(env: &mut Environment) {
-        if let Authority::Client { client_key, .. } = env.client {
-            let identifier = random::<XorName>();
-            let sd = unwrap_result!(StructuredData::new(0,
-                                                        identifier,
-                                                        0,
-                                                        vec![],
-                                                        vec![client_key],
-                                                        vec![],
-                                                        None));
-            let msg_id = MessageId::new();
-            let data = Data::Structured(sd);
-            let request = RequestMessage {
-                src: env.client.clone(),
-                dst: env.our_authority.clone(),
-                content: RequestContent::Put(data.clone(), msg_id),
-            };
-
-            assert!(env.maid_manager
-                       .handle_put(&request, &data, &msg_id)
-                       .is_ok());
-        };
-    }
-
-    fn get_close_node(env: &Environment) -> XorName {
-        let mut name = random::<XorName>();
-
-        loop {
-            if let Ok(Some(_)) = env.routing.close_group(name) {
-                return name;
-            } else {
-                name = random::<XorName>();
-            }
-        }
-    }
-
-    fn lose_close_node(env: &Environment) -> XorName {
-        let our_name = if let Ok(ref name) = env.routing.name() {
-            *name
-        } else {
-            unreachable!()
-        };
-        if let Ok(Some(close_group)) = env.routing.close_group(*env.our_authority.name()) {
-            let mut rng = thread_rng();
-            let range = Range::new(0, close_group.len());
-            assert_eq!(close_group.len(), GROUP_SIZE);
-            loop {
-                let index = range.ind_sample(&mut rng);
-                if close_group[index] != our_name {
-                    return close_group[index];
-                }
-            }
-        } else {
-            unreachable!()
-        }
-    }
-
 
     #[test]
     fn handle_put_without_account() {
-        let mut env = environment_setup();
+        let network = Network::new();
+        let node_count = TEST_NET_SIZE;
+        let mut nodes = test_node::create_nodes(&network, node_count, None);
+        let config = mock_crust::Config::with_contacts(&[nodes[0].endpoint()]);
+        let mut client = TestClient::new(&network, Some(config));
 
-        // Try with valid ImmutableData before account is created
-        let immutable_data = ImmutableData::new(generate_random_vec_u8(1024));
-        let msg_id = MessageId::new();
-        let data = Data::Immutable(immutable_data);
-        let valid_request = RequestMessage {
-            src: env.client.clone(),
-            dst: env.our_authority.clone(),
-            content: RequestContent::Put(data.clone(), msg_id),
-        };
+        client.ensure_connected(&mut nodes);
 
-        if let Err(InternalError::ClientMutation(MutationError::NoSuchAccount)) =
-               env.maid_manager.handle_put(&valid_request, &data, &msg_id) {
-        } else {
-            unreachable!()
-        }
-
-        let put_requests = env.routing.put_requests_given();
-
-        assert!(put_requests.is_empty());
-
-        let put_failures = env.routing.put_failures_given();
-
-        assert_eq!(put_failures.len(), 1);
-        assert_eq!(put_failures[0].src, env.our_authority);
-        assert_eq!(put_failures[0].dst, env.client);
-
-        if let ResponseContent::PutFailure { ref id, ref request, ref external_error_indicator } =
-               put_failures[0].content {
-            assert_eq!(*id, msg_id);
-            assert_eq!(*request, valid_request);
-            if let MutationError::NoSuchAccount =
-                   unwrap_result!(serialisation::deserialise(external_error_indicator)) {
-            } else {
-                unreachable!()
-            }
-        } else {
-            unreachable!()
-        }
+        let immutable_data = ImmutableData::new(test_utils::generate_random_vec_u8(1024));
+        client.put(Data::Immutable(immutable_data));
+        let _ = poll::nodes_and_client(&mut nodes, &mut client);
+        let count = nodes.iter()
+                         .filter(|node| node.get_maid_manager_put_count(client.name()).is_some())
+                         .count();
+        assert!(0 == count, "put_count {} found with {} nodes", count, count);
     }
 
     #[test]
     fn handle_put_with_account() {
-        let mut env = environment_setup();
-        create_account(&mut env);
+        let network = Network::new();
+        let node_count = TEST_NET_SIZE;
+        let mut nodes = test_node::create_nodes(&network, node_count, None);
+        let config = mock_crust::Config::with_contacts(&[nodes[0].endpoint()]);
+        let mut client = TestClient::new(&network, Some(config));
 
-        let immutable_data = ImmutableData::new(generate_random_vec_u8(1024));
-        let msg_id = MessageId::new();
-        let data = Data::Immutable(immutable_data.clone());
-        let valid_request = RequestMessage {
-            src: env.client.clone(),
-            dst: env.our_authority.clone(),
-            content: RequestContent::Put(data.clone(), msg_id),
-        };
+        client.ensure_connected(&mut nodes);
+        client.create_account(&mut nodes);
 
-        assert!(env.maid_manager
-                   .handle_put(&valid_request, &data, &msg_id)
-                   .is_ok());
-
-        let put_failures = env.routing.put_failures_given();
-        assert!(put_failures.is_empty());
-
-        let put_requests = env.routing.put_requests_given();
-
-        // put_requests[0] - account creation.
-        assert_eq!(put_requests.len(), 2);
-        assert_eq!(put_requests[1].src, env.our_authority);
-        assert_eq!(put_requests[1].dst,
-                   Authority::NaeManager(immutable_data.name()));
-
-        if let RequestContent::Put(Data::Immutable(ref data), ref id) = put_requests[1].content {
-            assert_eq!(*data, immutable_data);
-            assert_eq!(*id, msg_id);
-        } else {
-            unreachable!()
-        }
+        let immutable_data = ImmutableData::new(test_utils::generate_random_vec_u8(1024));
+        client.put(Data::Immutable(immutable_data.clone()));
+        let _ = poll::nodes_and_client(&mut nodes, &mut client);
+        let count = nodes.iter()
+                         .filter(|node| node.get_maid_manager_put_count(client.name()).is_some())
+                         .count();
+        assert!(GROUP_SIZE == count,
+                "clinet account {} found on {} nodes",
+                count,
+                count);
+        let mut stored_immutable = Vec::new();
+        stored_immutable.push(Data::Immutable(immutable_data));
+        mock_crust_detail::check_data(stored_immutable, &nodes);
     }
 
     #[test]
+    #[should_panic] // TODO Look at using std::panic::catch_unwind (1.9)
     fn invalid_put_for_previously_created_account() {
-        let mut env = environment_setup();
-        create_account(&mut env);
+        let network = Network::new();
+        let node_count = TEST_NET_SIZE;
+        let mut nodes = test_node::create_nodes(&network, node_count, None);
+        let config = mock_crust::Config::with_contacts(&[nodes[0].endpoint()]);
+        let mut client = TestClient::new(&network, Some(config));
 
-        let immutable_data = ImmutableData::new(generate_random_vec_u8(1024));
-        let mut msg_id = MessageId::new();
-        let data = Data::Immutable(immutable_data.clone());
-        let mut valid_request = RequestMessage {
-            src: env.client.clone(),
-            dst: env.our_authority.clone(),
-            content: RequestContent::Put(data.clone(), msg_id),
-        };
+        client.ensure_connected(&mut nodes);
+        client.create_account(&mut nodes);
+        client.create_account(&mut nodes);
+    }
 
-        assert!(env.maid_manager
-                   .handle_put(&valid_request, &data, &msg_id)
-                   .is_ok());
+    #[test]
+    fn maid_manager_churn() {
+        let network = Network::new();
+        let node_count = 15;
+        let mut nodes = test_node::create_nodes(&network, node_count, None);
+        let config = mock_crust::Config::with_contacts(&[nodes[0].endpoint()]);
+        let mut client = TestClient::new(&network, Some(config));
 
-        let mut put_failures = env.routing.put_failures_given();
-        assert!(put_failures.is_empty());
+        client.ensure_connected(&mut nodes);
+        client.create_account(&mut nodes);
 
-        let put_requests = env.routing.put_requests_given();
+        let mut rng = thread_rng();
 
-        assert_eq!(put_requests.len(), 2);
-        assert_eq!(put_requests[1].src, env.our_authority);
-        assert_eq!(put_requests[1].dst,
-                   Authority::NaeManager(immutable_data.name()));
+        let mut put_count = 1; // Login packet.
+        let full_id = client.full_id().clone();
 
-        if let RequestContent::Put(Data::Immutable(ref data), ref id) = put_requests[1].content {
-            assert_eq!(*data, immutable_data);
-            assert_eq!(*id, msg_id);
-        } else {
-            unreachable!()
-        }
-
-        let client_key = if let Authority::Client { client_key, .. } = env.client {
-            client_key
-        } else {
-            unreachable!()
-        };
-
-        let identifier = random::<XorName>();
-        let sd = unwrap_result!(StructuredData::new(0,
-                                                    identifier,
-                                                    0,
-                                                    vec![],
-                                                    vec![client_key],
-                                                    vec![],
-                                                    None));
-        msg_id = MessageId::new();
-        let sd_data = Data::Structured(sd);
-        valid_request = RequestMessage {
-            src: env.client.clone(),
-            dst: env.our_authority.clone(),
-            content: RequestContent::Put(sd_data.clone(), msg_id),
-        };
-
-        if let Err(InternalError::ClientMutation(MutationError::AccountExists)) =
-               env.maid_manager
-                  .handle_put(&valid_request, &sd_data, &msg_id) {
-        } else {
-            unreachable!()
-        }
-
-        put_failures = env.routing.put_failures_given();
-
-        assert_eq!(put_failures.len(), 1);
-        assert_eq!(put_failures[0].src, env.our_authority);
-        assert_eq!(put_failures[0].dst, env.client);
-
-        if let ResponseContent::PutFailure { ref id, ref request, ref external_error_indicator } =
-               put_failures[0].content {
-            assert_eq!(*id, msg_id);
-            assert_eq!(*request, valid_request);
-            if let MutationError::AccountExists =
-                   unwrap_result!(serialisation::deserialise(external_error_indicator)) {} else {
-                unreachable!()
+        for i in 0..10 {
+            for data in (0..4).map(|_| {
+                Data::Structured(test_utils::random_structured_data(100000, &full_id))
+            }) {
+                client.put(data.clone());
+                put_count += 1;
             }
-        } else {
-            unreachable!()
-        }
-    }
-
-    #[test]
-    fn handle_put_success() {
-        let mut env = environment_setup();
-        create_account(&mut env);
-
-        let immutable_data = ImmutableData::new(generate_random_vec_u8(1024));
-        let mut msg_id = MessageId::new();
-        let data = Data::Immutable(immutable_data.clone());
-        let valid_request = RequestMessage {
-            src: env.client.clone(),
-            dst: env.our_authority.clone(),
-            content: RequestContent::Put(data.clone(), msg_id),
-        };
-
-        assert!(env.maid_manager
-                   .handle_put(&valid_request, &data, &msg_id)
-                   .is_ok());
-
-        let put_failures = env.routing.put_failures_given();
-        assert!(put_failures.is_empty());
-
-        let put_requests = env.routing.put_requests_given();
-
-        assert_eq!(put_requests.len(), 2);
-        assert_eq!(put_requests[1].src, env.our_authority);
-        assert_eq!(put_requests[1].dst,
-                   Authority::NaeManager(immutable_data.name()));
-
-        let data = if let RequestContent::Put(Data::Immutable(ref data), ref id) =
-                          put_requests[1].content {
-            assert_eq!(*data, immutable_data);
-            assert_eq!(*id, msg_id);
-            data
-        } else {
-            unreachable!()
-        };
-
-        // Valid case.
-        assert!(env.maid_manager
-                   .handle_put_success(&data.identifier(), &msg_id)
-                   .is_ok());
-
-        let put_successes = env.routing.put_successes_given();
-
-        assert_eq!(put_successes.len(), 1);
-        assert_eq!(put_successes[0].src, env.our_authority);
-        assert_eq!(put_successes[0].dst, env.client);
-
-        if let ResponseContent::PutSuccess(ref name, ref id) = put_successes[0].content {
-            assert_eq!(*id, msg_id);
-            assert_eq!(*name, data.identifier());
-        } else {
-            unreachable!()
-        }
-
-        // Invalid case.
-        msg_id = MessageId::new();
-
-        if let Err(InternalError::FailedToFindCachedRequest(id)) =
-               env.maid_manager.handle_put_success(&data.identifier(), &msg_id) {
-            assert_eq!(msg_id, id);
-        } else {
-            unreachable!()
-        }
-    }
-
-    #[test]
-    fn handle_put_failure() {
-        let mut env = environment_setup();
-        create_account(&mut env);
-
-        let client_key = if let Authority::Client { client_key, .. } = env.client {
-            client_key
-        } else {
-            unreachable!()
-        };
-        let identifier = random::<XorName>();
-        let sd = unwrap_result!(StructuredData::new(1,
-                                                    identifier,
-                                                    0,
-                                                    vec![],
-                                                    vec![client_key],
-                                                    vec![],
-                                                    None));
-        let mut msg_id = MessageId::new();
-        let data = Data::Structured(sd.clone());
-        let valid_request = RequestMessage {
-            src: env.client.clone(),
-            dst: env.our_authority.clone(),
-            content: RequestContent::Put(data.clone(), msg_id),
-        };
-
-        assert!(env.maid_manager
-                   .handle_put(&valid_request, &data, &msg_id)
-                   .is_ok());
-
-        let mut put_failures = env.routing.put_failures_given();
-        assert!(put_failures.is_empty());
-
-        let put_requests = env.routing.put_requests_given();
-
-        assert_eq!(put_requests.len(), 2);
-        assert_eq!(put_requests[1].src, env.our_authority);
-        assert_eq!(put_requests[1].dst, Authority::NaeManager(sd.name()));
-
-        if let RequestContent::Put(Data::Structured(ref data), ref id) = put_requests[1].content {
-            assert_eq!(*data, sd);
-            assert_eq!(*id, msg_id);
-        } else {
-            unreachable!()
-        }
-
-        // Valid case.
-        let error = MutationError::NoSuchData;
-        if let Ok(error_indicator) = serialisation::serialise(&error) {
-            assert!(env.maid_manager
-                       .handle_put_failure(&msg_id, &error_indicator[..])
-                       .is_ok());
-        } else {
-            unreachable!()
-        }
-
-        put_failures = env.routing.put_failures_given();
-
-        assert_eq!(put_failures.len(), 1);
-        assert_eq!(put_failures[0].src, env.our_authority);
-        assert_eq!(put_failures[0].dst, env.client);
-
-        if let ResponseContent::PutFailure { ref id, ref request, ref external_error_indicator } =
-               put_failures[0].content {
-            assert_eq!(*id, msg_id);
-            assert_eq!(*request, valid_request);
-            if let Ok(error_indicator) = serialisation::serialise(&error) {
-                assert_eq!(*external_error_indicator, error_indicator);
+            trace!("Churning on {} nodes, iteration {}", nodes.len(), i);
+            if nodes.len() <= GROUP_SIZE + 2 || random() {
+                let index = Range::new(1, nodes.len()).ind_sample(&mut rng);
+                trace!("Adding node with bootstrap node {}.", index);
+                test_node::add_node(&network, &mut nodes, index);
             } else {
-                unreachable!()
-            }
-        } else {
-            unreachable!()
-        }
-
-        // Invalid case.
-        msg_id = MessageId::new();
-        if let Ok(error_indicator) = serialisation::serialise(&error) {
-            if let Err(InternalError::FailedToFindCachedRequest(id)) =
-                   env.maid_manager.handle_put_failure(&msg_id, &error_indicator[..]) {
-                assert_eq!(msg_id, id);
-            } else {
-                unreachable!()
-            }
-        } else {
-            unreachable!()
-        }
-    }
-
-    // #[test]
-    // fn network_full() {
-    //     let mut env = environment_setup();
-    //     create_account(&mut env);
-
-    //     let immutable_data = ImmutableData::new(generate_random_vec_u8(1024));
-    //     let msg_id = MessageId::new();
-    //     let data = Data::Immutable(immutable_data.clone());
-    //     let valid_request = RequestMessage {
-    //         src: env.client.clone(),
-    //         dst: env.our_authority.clone(),
-    //         content: RequestContent::Put(data.clone(), msg_id),
-    //     };
-
-    // let mut full_pmid_nodes = HashSet::new();
-
-    //     if let Ok(Some(close_group)) = env.routing.close_group(utils::client_name(&env.client)) {
-    //         full_pmid_nodes = close_group.iter()
-    //                                      .take(close_group.len() / 2)
-    //                                      .cloned()
-    //                                      .collect::<HashSet<XorName>>();
-    //     }
-
-    //     assert!(env.maid_manager
-    //                .handle_put(&env.routing, &valid_request, &data, &msg_id)
-    //                .is_ok());
-
-    // let put_failures = env.routing.put_failures_given();
-
-    //     assert_eq!(put_failures.len(), 1);
-    //     assert_eq!(put_failures[0].src, env.our_authority);
-    //     assert_eq!(put_failures[0].dst, env.client);
-
-    //     if let ResponseContent::PutFailure { ref id, ref request, ref external_error_indicator } =
-    //            put_failures[0].content {
-    //         assert_eq!(*id, msg_id);
-    //         assert_eq!(*request, valid_request);
-    //         if let Ok(error_indicator) = serialisation::serialise(&MutationError::NetworkFull) {
-    //             assert_eq!(*external_error_indicator, error_indicator);
-    //         } else {
-    //             unreachable!()
-    //         }
-    //     } else {
-    //         unreachable!()
-    //     }
-    // }
-
-    #[test]
-    fn churn_refresh() {
-        let mut env = environment_setup();
-        create_account(&mut env);
-
-        let mut refresh_count = 1;
-        let mut refresh_requests = env.routing.refresh_requests_given();
-        assert_eq!(refresh_requests.len(), refresh_count);
-        assert_eq!(refresh_requests[0].src, env.our_authority);
-        assert_eq!(refresh_requests[0].dst, env.our_authority);
-
-        env.routing.node_added_event(get_close_node(&env));
-        env.maid_manager.handle_node_added(&random::<XorName>());
-
-        refresh_requests = env.routing.refresh_requests_given();
-
-        if let Ok(Some(_)) = env.routing.close_group(utils::client_name(&env.client)) {
-            assert_eq!(refresh_requests.len(), refresh_count + 1);
-            assert_eq!(refresh_requests[refresh_count].src, env.our_authority);
-            assert_eq!(refresh_requests[refresh_count].dst, env.our_authority);
-            refresh_count += 1;
-
-            if let RequestContent::Refresh(ref serialised_refresh, _) = refresh_requests[0]
-                                                                            .content {
-                if let Ok(refresh) = serialisation::deserialise(&serialised_refresh) {
-                    let refresh: Refresh = refresh;
-                    assert_eq!(refresh.0, utils::client_name(&env.client));
-                } else {
-                    unreachable!()
+                let number = Range::new(1, 4).ind_sample(&mut rng);
+                trace!("Removing {} node(s).", number);
+                for _ in 0..number {
+                    let node_index = Range::new(1, nodes.len()).ind_sample(&mut rng);
+                    test_node::drop_node(&mut nodes, node_index);
                 }
-            } else {
-                unreachable!()
             }
-        } else {
-            assert_eq!(refresh_requests.len(), refresh_count);
-        }
-
-        let node_lost = lose_close_node(&env);
-        env.routing.node_lost_event(node_lost.clone());
-        env.maid_manager.handle_node_lost(&node_lost);
-
-        refresh_requests = env.routing.refresh_requests_given();
-
-        if let Ok(Some(_)) = env.routing.close_group(utils::client_name(&env.client)) {
-            assert_eq!(refresh_requests.len(), refresh_count + 1);
-            assert_eq!(refresh_requests[refresh_count].src, env.our_authority);
-            assert_eq!(refresh_requests[refresh_count].dst, env.our_authority);
-            // refresh_count += 1;
-
-            if let RequestContent::Refresh(ref serialised_refresh, _) =
-                   refresh_requests[refresh_count].content {
-                if let Ok(refresh) = serialisation::deserialise(&serialised_refresh) {
-                    let refresh: Refresh = refresh;
-                    assert_eq!(refresh.0, utils::client_name(&env.client));
-                } else {
-                    unreachable!()
-                }
-            } else {
-                unreachable!()
-            }
-        } else {
-            assert_eq!(refresh_requests.len(), refresh_count);
+            let _ = poll::nodes_and_client(&mut nodes, &mut client);
+            let count = nodes.iter()
+                             .filter(|node| {
+                                 match node.get_maid_manager_put_count(client.name()) {
+                                     None => false,
+                                     Some(count) => count == put_count,
+                                 }
+                             })
+                             .count();
+            assert!(GROUP_SIZE - 3 <= count,
+                    "put_count {} only found with {} nodes",
+                    put_count,
+                    count);
         }
     }
 }
